+++ webpack.config.js
@@ -13,7 +13,6 @@ module.exports = {
     products: asset('js/products.js'),
     pages: asset('js/pages.js'),
     'zod-menu': asset('js/partials/zod-menu.js'),
-    'product-card-marketplace': asset('js/product-card-marketplace.js'),
     'product-runtime-compat': asset('js/product-runtime-compat.js'),
     'digital-files': asset('js/partials/digital-files.js')
   },